Refactor plot generator into modular components
