+++ src/Tools/Plot_Generator/__init__.py
@@ -1,5 +1,7 @@
 """Plot generator tool for creating SNR/BCA line plots."""
 
-from .plot_generator import PlotGeneratorWindow, main
+from .gui import PlotGeneratorWindow
+from .worker import _Worker
+from .plot_generator import main
 
-__all__ = ["PlotGeneratorWindow", "main"]
+__all__ = ["_Worker", "PlotGeneratorWindow", "main"]
